fix(BA-4994): consolidate agent tests into test_agent_api.py
Moved compound filter and empty result tests from test_agent_operations.py
into test_agent_api.py::TestSearchAgents class. Removed duplicate permission
test and deleted test_agent_operations.py. Also cleaned up fix-report.md.

Changes:
- Added test_compound_filters_status_and_resource_group to TestSearchAgents
- Added test_empty_result_nonexistent_resource_group to TestSearchAgents
- Deleted tests/component/agent_api/test_agent_operations.py
- Deleted fix-report.md from worktree root
- Updated changelog to accurately describe added tests
