+++ changes/9911.test.md
@@ -1,1 +1,1 @@
-Add component tests for agent search operations with compound filters and empty results - complements existing agent API tests with additional test coverage for filter combinations and edge cases
+Add component tests for agent search operations - compound filter scenarios (status + resource_group intersection) and empty result handling for nonexistent resource groups
